// src/app/models/revenueModel.ts
export interface Revenue {
    productGroup?: any;
    fullProductGroup?: any;
    productGroupPourCent?: any;
    revenueProductGroup?: any;
    reportNo?: any;
}
export interface INoSalesAtAllByCategory {
    prodNum?: any;
    prodName?: any;
}
export interface IProductByReportNo {
    prodName?: any;
    quantity?: any;
    revenueProductGroup?: any;
}
export interface ITotalRevenueAndQuality {
    quanlity?: any;
    revenueProductGroup?: any;
}